Bugfix on sprint.c fixing weird bugs in ar
If you add a few fprint in ar and sprint.c,

Before:

~/kencc/ar/o.out vu lib.a foo.8
HERE lib.a 1 foo.8
objsym:foo, symdefsize=9
sprint: 5db96d94, len = 1073741824, fmt = %-12ld
sprint: 5db96da0, len = 1073741824, fmt = %-6d
sprint: 5db96da6, len = 1073741824, fmt = %-6d
sprint: 5db96dac, len = 1073741824, fmt = %-8lo
sprint: 5db96db4, len = 1073741824, fmt = %-10lld
ar: creating lib.a
sprint: ffa44dbc, len = 1073741824, fmt = %-12ld
sprint: ffa44dc8, len = 1073741824, fmt = %-6d
sprint: ffa44dce, len = 1073741824, fmt = %-6d
rl sprint 0 0 0 0 0 0
=======> PAD: didn't do the sprint!
sprint: ffa44dd4, len = 1073741824, fmt = %-8o
sprint: ffa44ddc, len = 1073741824, fmt = %-10ld
rl: fd=3, __.SYMDEF       , len = 10
wrsym: foo len = 4
arwrite: fd = 3, member = foo.8           1750666821  0     0     644     126       `
~
a - foo.8

master/home/pad/principia $ rm lib.a
/usr/bin/rm: remove regular file 'lib.a'? y
master/home/pad/principia $ ~/kencc/ar/o.out vu lib.a foo.8
HERE lib.a 1 foo.8
objsym:foo, symdefsize=9
sprint: 58578d94, len = 1073741824, fmt = %-12ld
sprint: 58578da0, len = 1073741824, fmt = %-6d
sprint: 58578da6, len = 1073741824, fmt = %-6d
sprint: 58578dac, len = 1073741824, fmt = %-8lo
sprint: 58578db4, len = 1073741824, fmt = %-10lld
ar: creating lib.a
sprint: ffaca0fc, len = 5463811, fmt = %-12ld
sprint: ffaca108, len = 5463799, fmt = %-6d
sprint: ffaca10e, len = 5463793, fmt = %-6d
rl sprint 52 50 32 32 32 32
=========> PAD this time it did it!
sprint: ffaca114, len = 5463787, fmt = %-8o
sprint: ffaca11c, len = 5463779, fmt = %-10ld
rl: fd=3, __.SYMDEF       1750674848  42    42    644     10        `
, len = 10
wrsym: foo len = 4
arwrite: fd = 3, member = foo.8           1750666821  0     0     644     126       `
~
a - foo.8

+++ lib9/sprint.c
@@ -25,8 +25,12 @@ sprint(char *buf, char *fmt, ...)
 	 * the stack might be near the top of memory, so
 	 * we must be sure not to overflow a 32-bit pointer.
 	 */
-	if(buf+len < buf)
-		len = -(uint)buf-1;
+        //BUGFIX: without the (uintptr) cast, sprint is actually
+        // not working when buf is a pointer in the stack
+        // which has very bad implications (e.g., 'ar' not working).
+        // I still don't really understand the fix though. Ugly C.
+	if((uintptr)buf+len < (uintptr)buf)
+		len = (uint)-(uintptr)buf-1;
 
 	va_start(args, fmt);
 	n = vsnprint(buf, len, fmt, args);
